Library update

+++ main/menu.c
@@ -4,8 +4,8 @@
 #include "watchy.h"
 #include "gfx.h"
 
-uint8_t
-menu_show (uint8_t menu, uint8_t buttons)
+uint32_t
+menu_show (uint32_t menu, uint8_t buttons)
 {
 
    return 0;
