+++ main/menu.h
@@ -1,3 +1,3 @@
 // menu
 
-uint8_t menu_show (uint8_t menu, uint8_t button);
+uint32_t menu_show (uint32_t menu, uint8_t button);
